Fix datatypes
int -> long for money

+++ TornJson/UserData/CityBank.cs
@@ -22,7 +22,7 @@ namespace TornJson.UserData
 {
     public class CityBank
     {
-        [JsonProperty("amount")] public int Amount { get; set; }
+        [JsonProperty("amount")] public long Amount { get; set; }
 
         [JsonProperty("time_left")] public int TimeLeft { get; set; }
     }
